Added navbar, gallery.
